refactor: 콘솔 입력 방식 ConsoleInputService로 분리
